feat: OrderStatus에 따른 삭제 기능 구현

+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/orderinfo/service/OrderInfoService.java
@@ -63,7 +63,14 @@ public class OrderInfoService {
         orderInfo.setMemberAddress(memberAddress);
     }
 
-    public void deleteOrderInfo(OrderInfo orderInfo) {
+    @Transactional
+    public void deleteOrderInfo(long orderId) {
+        OrderInfo orderInfo = getOrderById(orderId);
+
+        if (!orderInfo.getOrderStatus().equals(OrderStatus.ORDERED)) {
+            throw new ServiceException(HttpStatus.BAD_REQUEST, "해당 주문은 삭제할 수 없습니다.");
+        }
+
         orderInfoRepository.delete(orderInfo);
     }
 }
